admin login functionality
This commit creates a basic view with working
backend functionality for an admin user to log
into the system and access their own dashboard.

Admin credentials are protected by encryption
throughout the code. Furthermore, the url used
to access the admin login page is a random string,
blocked by Google search results to make it harder
to find for regular users.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>