upload and verify card image
This commit adds the code for uploading an image
of the student card during the registration process
and cross referencing the image with the form submitted
to make sure that values such as name and student number
match.

This commit does not yet store in the image in the DB
and front end has not been refined.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>