add liveness detection test
When verifying the person's face ID on voting, new
functionality has been added to test that the face
detected is coming from a real person and not a photo.

Firstly, in order to take the picture, the user must
blink firmly for the camera.

Secondly, the user must stay relatively still in order
to ensure a photo is not being waved in front of the
camera to trick the API.

Lastly, a frame has been added to help the user indicate
whether their face is being detected or not.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>